Extracted "overwriteFiles" out of "extract-addon" into separate module
