+++ lib/utils/extract-addon.js
@@ -1,6 +1,8 @@
-var Promise = require('../ext/promise');
 var fs = require('fs-extra');
 var path = require('path');
+
+var Promise = require('../ext/promise');
+var overwriteFiles = require('./overwriteFiles');
 
 var ensureDir = Promise.denodeify(fs.ensureDir);
 var copy = Promise.denodeify(fs.copy);
@@ -31,18 +33,6 @@ function replaceInFile(file, placeholder, value) {
 function replaceInFiles(files, placeholder, value) {
   return Promise.map(files, function(file) {
     return replaceInFile(file, placeholder, value);
-  });
-}
-
-function copyFileIfExists(oldFile, newFile) {
-  return new Promise(function(resolve) {
-    copy(oldFile, newFile, { clobber: true }).finally(resolve);
-  });
-}
-
-function overwriteFiles(fileMappings) {
-  return Promise.map(fileMappings, function(fileMapping) {
-    return copyFileIfExists(fileMapping.oldPath, fileMapping.newPath);
   });
 }
